backend/requirements/base.txt
# Core Django and API frameworks
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database and migrations
psycopg2-binary==2.9.8
django-environ==0.11.2

# API documentation and CORS
drf-spectacular==0.26.5
django-cors-headers==4.3.1

# Filtering and search
django-filter==23.4

# Image handling
Pillow==10.1.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Email and notifications
celery==5.3.4

# Development and debugging
django-debug-toolbar==4.2.0

# Testing
factory-boy==3.3.0
pytest-django==4.7.0

# File uploads and storage
django-storages==1.14.2
boto3==1.34.0

# Security
cryptography==41.0.7